Reshuffle most of program
