+++ src/Promitor.Agents.Scraper/Promitor.Agents.Scraper.csproj
@@ -37,10 +37,7 @@
     <PackageReference Include="AutoMapper" Version="9.0.0" />
     <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="7.0.0" />
     <PackageReference Include="CronScheduler.AspNetCore" Version="3.0.0" />
-    <PackageReference Include="NetEscapades.Configuration.Yaml" Version="2.0.1" />
     <PackageReference Include="Prometheus.Client.AspNetCore" Version="3.1.0" />
-    <PackageReference Include="Serilog.AspNetCore" Version="3.2.0" />
-    <PackageReference Include="Serilog.Sinks.ApplicationInsights" Version="3.1.0" />
     <PackageReference Include="System.Net.Http" Version="4.3.4" />
     <PackageReference Include="YamlDotNet" Version="8.1.1" />
   </ItemGroup>
